+++ smol/moca/sampler/kernel.py
@@ -313,7 +313,7 @@ class ThermalKernel(MCKernel):
             Trace
         """
         trace = super().compute_initial_trace(occupancy)
-        trace.temperature = self.trace.temperature
+        trace.temperature = np.array([self.trace.temperature])
         return trace
 
     def set_aux_state(self, occupancies, *args, **kwargs):